Adding more models to databse
