+++ src/app.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { ApiModule } from './for-react-api/api.module';
+import { ApiModule } from './react-api/api.module';
 
 @Module({
   imports: [ApiModule],
