--- src/for-react-api/api.controller.ts
@@ -1,13 +1,0 @@
-import { Controller, Get } from '@nestjs/common';
-import { ApiService } from './api.service';
-import { Observable } from 'rxjs';
-
-@Controller()
-export class ApiController {
-  constructor(private apiService: ApiService) {}
-
-  @Get('site-info')
-  getSiteInfo() {
-    return this.apiService.getSiteInfo();
-  }
-}
