+++ src/react-api/api.service.ts
@@ -11,4 +11,20 @@ export class ApiService {
   getSiteInfo() {
     return this.prisma.siteInfo.findMany();
   }
+
+  getCategories() {
+    return this.prisma.navbarCategories.findMany();
+  }
+
+  getCategoriesAccounts() {
+    return this.prisma.navbarCategories.findMany();
+  }
+
+  getSlobyMenuBar() {
+    return this.prisma.dashboardMenuBar.findMany();
+  }
+
+  getSlobyTools() {
+    return this.prisma.slobyTools.findMany();
+  }
 }
